fix: brew env init before brew install
- add git, vim, install tools step for command -s option
- fix brew install command issue
- remove homebrew in default config file
- remove universal-ctags install by brew
- emerge add -n flag
- fix warning message function
- add git commit template
- add homebrew environments in variables file

Closes #74

+++ bin/deps/commands.sh
@@ -97,16 +97,22 @@ pkg_install()
 brew_install()
 {
 	if ! command_exists "brew"; then
-		error_exit "you must install Homebrew first!"
-	else
-		local -a install_cmd=(brew install)
-		
-		for install_arg in "$@"
-		do
-			install_cmd+=("$install_arg")
-		done
-		execute "${install_cmd[@]}"
+        # homebrew has already installed, but not run brew's settings
+        if [ -f "/home/linuxbrew/.linuxbrew/bin/brew" ]; then
+            echo "homebrew already installed, loading..."
+            eval "$HOMEBREW_SETTINGS"
+            eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"
+        else
+            error_exit "you must install Homebrew first!"
+        fi
 	fi
+    local -a install_cmd=(brew install)
+    
+    for install_arg in "$@"
+    do
+        install_cmd+=("$install_arg")
+    done
+    execute "${install_cmd[@]}"
 }
 
 echo_commands()
@@ -136,7 +142,7 @@ gentoo_setup_portage_license()
 
 setup_gentoo()
 {
-    SYS_INSTALL_PKG_CMD=("emerge" "-v")
+    SYS_INSTALL_PKG_CMD=("emerge" "-vn")
     SYS_UPGRADE_PKG_CMD=("emerge" "--update" "--deep" "--change-use")
     SYS_UPDATE_CMD=("emerge" "--sync")
     SYS_UPGRADE_CMD=("emerge" "-avuND" "@world")
@@ -154,8 +160,8 @@ setup_gentoo()
 
 setup_manjaro()
 {
-    SYS_INSTALL_PKG_CMD=("pacman" "-Sy")
-    SYS_UPGRADE_PKG_CMD=("pacman" "-Sy")
+    SYS_INSTALL_PKG_CMD=("pacman" "-Sy --needed")
+    SYS_UPGRADE_PKG_CMD=("pacman" "-Syu --needed")
     SYS_UPDATE_CMD=("pacman" "-Syy")
     SYS_UPGRADE_CMD=("pacman" "-Syu")
 }
